requirements.txt
numpy>=1.11.2
six>=1.10.0
gunicorn
click==6.7
Cython==0.28.4
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
scikit-surprise
pandas
WTForms==2.1
